add bitmap() device, based on dev2bitmap() add destdir argument to {install,update}.packages
git-svn-id: https://svn.r-project.org/R/trunk@9005 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/base/R/unix/update.packages.R
@@ -1,6 +1,6 @@
 install.packages <- function(pkgs, lib, CRAN=getOption("CRAN"),
                              contriburl=contrib.url(CRAN),
-                             method="auto", available=NULL)
+                             method="auto", available=NULL, destdir=NULL)
 {
     if(missing(lib) || is.null(lib)) {
         lib <- .lib.loc[1]
@@ -8,8 +8,10 @@ install.packages <- function(pkgs, lib, CRAN=getOption("CRAN"),
     }
     localcran <- length(grep("^file:", contriburl)) > 0
     if(!localcran) {
-        tmpd <- tempfile("Rinstdir")
-        dir.create(tmpd)
+        if (is.null(destdir)){
+            tmpd <- tempfile("Rinstdir")
+            dir.create(tmpd)
+        } else tmpd <- destdir
     }
 
     foundpkgs <- download.packages(pkgs, destdir=tmpd,
@@ -40,7 +42,7 @@ install.packages <- function(pkgs, lib, CRAN=getOption("CRAN"),
             }
         }
         cat("\n")
-        if(!localcran){
+        if(!localcran && is.null(destdir)){
             answer <- substr(readline("Delete downloaded files (y/N)? "), 1, 1)
             if(answer == "y" | answer == "Y")
                 unlink(tmpd)
@@ -79,7 +81,7 @@ download.packages <- function(pkgs, destdir, available=NULL,
             else{
                 url <- paste(contriburl, fn, sep="/")
                 destfile <- file.path(destdir, fn)
-                
+
                 if(download.file(url, destfile, method) == 0)
                     retval <- rbind(retval, c(p, destfile))
                 else
